+++ egui_glium/src/lib.rs
@@ -97,8 +97,6 @@ pub use epi_backend::{run, NativeOptions};
 
 pub use egui_winit;
 
-use glium::glutin;
-
 // ----------------------------------------------------------------------------
 
 /// Convenience wrapper for using [`egui`] from a [`glium`] app.
@@ -125,11 +123,6 @@ impl EguiGlium {
     /// Note that egui uses `tab` to move focus between elements, so this will always return `true` for tabs.
     pub fn on_event(&mut self, event: &glium::glutin::event::WindowEvent<'_>) -> bool {
         self.egui_winit.on_event(&self.egui_ctx, event)
-    }
-
-    /// Is this a close event or a Cmd-Q/Alt-F4 keyboard command?
-    pub fn is_quit_event(&self, event: &glutin::event::WindowEvent<'_>) -> bool {
-        self.egui_winit.is_quit_event(event)
     }
 
     /// Returns `needs_repaint` and shapes to draw.