+++ egui_glium/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the `egui_glium` integration will be noted in this file.
 
 ## Unreleased
 * Simplify `EguiGlium` interface ([#871](https://github.com/emilk/egui/pull/871)).
+* Remove `EguiGlium::is_quit_event` ([#881](https://github.com/emilk/egui/pull/881)).
 
 
 ## 0.15.0 - 2021-10-24